+++ src/model/database/connectionNode.ts
@@ -18,12 +18,12 @@ import { UserGroup } from "./userGroup";
  */
 export class ConnectionNode extends Node implements CopyAble {
 
-    public iconPath: string = path.join(Constants.RES_PATH, "icon/server.png");
+    public iconPath: string|vscode.ThemeIcon = path.join(Constants.RES_PATH, "icon/server.png");
     public contextValue: string = ModelType.CONNECTION;
     constructor(readonly key: string, readonly parent: Node) {
         super(key)
         this.init(parent)
-        this.label = (this.usingSSH) ? `${this.ssh.host}@${this.ssh.port}` : `${this.host}@${this.port}`;
+        this.label = (this.usingSSH) ? `${this.ssh.host}@${this.ssh.port}` : `${this.host}@${this.instanceName?this.instanceName:this.port}`;
         this.cacheSelf()
         if (parent.name) {
             this.description = parent.name
@@ -31,7 +31,8 @@ export class ConnectionNode extends Node implements CopyAble {
         }
         if (this.disable) {
             this.collapsibleState = vscode.TreeItemCollapsibleState.None;
-            this.iconPath = path.join(Constants.RES_PATH, "icon/close.svg");
+            this.iconPath = Global.disableIcon;
+            this.label=this.label+" (closed)"
             return;
         }
         const lcp = ConnectionManager.activeNode;
@@ -70,13 +71,13 @@ export class ConnectionNode extends Node implements CopyAble {
                 const includeDatabaseArray = this.includeDatabases?.toLowerCase()?.split(",")
                 const usingInclude = this.includeDatabases && includeDatabaseArray && includeDatabaseArray.length >= 1;
                 const databaseNodes = databases.filter((db) => {
-                    if (usingInclude) {
-                        return includeDatabaseArray.indexOf((db.schema || db.Database).toLocaleLowerCase()) != -1;
+                    if (usingInclude  && !db.schema) {
+                        return includeDatabaseArray.indexOf(db.Database.toLocaleLowerCase()) != -1;
                     }
                     return true;
                 }).map<SchemaNode | CatalogNode>((database) => {
                     return hasCatalog ?
-                        new CatalogNode(database.db, this)
+                        new CatalogNode(database.Database, this)
                         : new SchemaNode(database.schema || database.Database, this);
                 });