+++ src/service/queryUnit.ts
@@ -15,7 +15,6 @@ import { Trans } from "~/common/trans";
 import { IConnection } from "./connect/connection";
 import { FieldInfo } from "@/common/typeDef";
 import { Util } from "@/common/util";
-import { utils } from "ssh2-streams";
 
 export class QueryUnit {
 
@@ -50,7 +49,7 @@ export class QueryUnit {
 
         let recordHistory = queryOption.recordHistory;
         if (!sql) {
-            sql = this.getSqlFromEditor(connectionNode);
+            sql = this.getSqlFromEditor(connectionNode,queryOption.runAll);
             recordHistory = true;
         }
         sql = sql.replace(/^\s*--.+/igm, '').trim();
@@ -89,31 +88,31 @@ export class QueryUnit {
                 if (recordHistory) {
                     vscode.commands.executeCommand(CodeCommand.RecordHistory, sql, costTime);
                 }
-                if (data.affectedRows) {
-                    QueryPage.send({ connection: connectionNode, type: MessageType.DML, queryOption, res: { sql, costTime, affectedRows: data.affectedRows } as DMLResponse });
-                    // vscode.commands.executeCommand(CommandKey.Refresh);
-                    return;
+
+                if (sql.match(/create (table|prcedure|FUNCTION|VIEW)/i)) {
+                    vscode.commands.executeCommand(CodeCommand.Refresh);
                 }
 
-                // unknow result
-                if (!Array.isArray(data)) {
-                    QueryPage.send({ connection: connectionNode, type: MessageType.MESSAGE_BLOCK, queryOption, res: { sql, costTime } as DMLResponse });
+                if (data.affectedRows) {
+                    QueryPage.send({ connection: connectionNode, type: MessageType.DML, queryOption, res: { sql, costTime, affectedRows: data.affectedRows } as DMLResponse });
                     return;
                 }
 
                 // query result
-                if(fields && Array.isArray(fields) && fields[0]!=null && fields[0].name!=undefined){
+                if (Array.isArray(fields) && fields[0] != null && fields[0].name != undefined) {
                     QueryPage.send({ connection: connectionNode, type: MessageType.DATA, queryOption, res: { sql, costTime, data, fields, total, pageSize: Global.getConfig(ConfigKey.DEFAULT_LIMIT) } as DataResponse });
                     return;
                 }
 
-                // mysql procedrue call result
-                const lastEle = data[data.length - 1]
-                if (data.length>2 &&  Util.is(lastEle,'ResultSetHeader') && Util.is(data[0],'TextRow')) {
-                    data = data[data.length - 2]
-                    fields = fields[fields.length - 2] as any as FieldInfo[]
-                    QueryPage.send({ connection: connectionNode, type: MessageType.DATA, queryOption, res: { sql, costTime, data, fields, total, pageSize: Global.getConfig(ConfigKey.DEFAULT_LIMIT) } as DataResponse });
-                    return;
+                if (Array.isArray(data)) {
+                    // mysql procedrue call result
+                    const lastEle = data[data.length - 1]
+                    if (data.length > 2 && Util.is(lastEle, 'ResultSetHeader') && Util.is(data[0], 'TextRow')) {
+                        data = data[data.length - 2]
+                        fields = fields[fields.length - 2] as any as FieldInfo[]
+                        QueryPage.send({ connection: connectionNode, type: MessageType.DATA, queryOption, res: { sql, costTime, data, fields, total, pageSize: Global.getConfig(ConfigKey.DEFAULT_LIMIT) } as DataResponse });
+                        return;
+                    }
                 }
 
                 QueryPage.send({ connection: connectionNode, type: MessageType.MESSAGE_BLOCK, queryOption, res: { sql, costTime } as DMLResponse });
@@ -146,12 +145,16 @@ export class QueryUnit {
 
     private static batchPattern = /\s+(TRIGGER|PROCEDURE|FUNCTION)\s+/ig;
 
-    private static getSqlFromEditor(connectionNode: Node): string {
+    private static getSqlFromEditor(connectionNode: Node,runAll:boolean): string {
         if (!vscode.window.activeTextEditor) {
             throw new Error("No SQL file selected!");
 
         }
         const activeTextEditor = vscode.window.activeTextEditor;
+        if(runAll){
+            return activeTextEditor.document.getText()
+        }
+
         const selection = activeTextEditor.selection;
         const newLocal = !selection.isEmpty ? activeTextEditor.document.getText(selection) :
             this.obtainSql(activeTextEditor, DelimiterHolder.get(connectionNode.getConnectId()));
@@ -219,4 +222,8 @@ export interface QueryOption {
     viewId?: any;
     split?: boolean;
     recordHistory?: boolean;
+    /**
+     * runAll if get sql from editor.
+     */
+    runAll?: boolean;
 }